Move nvbench-related stuff to its own header
