+++ benchmarks/hash_table/static_map/contains_bench.cu
@@ -17,6 +17,7 @@
 #include <defaults.hpp>
 #include <distribution.hpp>
 #include <key_generator.hpp>
+#include <utils.hpp>
 
 #include <cuco/static_map.cuh>
