+++ benchmarks/hash_table/static_map/find_bench.cu
@@ -16,6 +16,7 @@
 
 #include <defaults.hpp>
 #include <key_generator.hpp>
+#include <utils.hpp>
 
 #include <cuco/static_map.cuh>
